Added some logic

+++ mobile/src/test/java/com/eckerlin/todoq/models/TaskListTest.java
@@ -1,8 +1,5 @@
 package com.eckerlin.todoq.models;
 
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -17,18 +14,19 @@ public class TaskListTest {
 
     @Before
     public void setUp() throws IOException {
-        Context appContext = InstrumentationRegistry.getTargetContext();
-        File taskFile = new File(appContext.getFilesDir().getPath(), "task.list");
-
+        File taskFile = new File(".", "task.list");
+/*        if (taskFile.exists()) {
+            taskFile.delete();
+        }*/
         list = new TaskList(taskFile);
     }
 
     @After
     public void tearDown() throws IOException {
-        Context appContext = InstrumentationRegistry.getTargetContext();
-        File taskFile = new File(appContext.getFilesDir().getPath(), "task.list");
-
-        taskFile.delete();
+        list.close();
+        File taskFile = new File(".", "task.list");
+        if (taskFile.exists())
+            taskFile.delete();
     }
 
     @Test
@@ -50,11 +48,43 @@ public class TaskListTest {
 
         list.push(t1);
         list.push(t2);
+        assertEquals(list.peek(), list.peek());
+    }
+
+    @Test
+    public void peekAndPopShouldReturnTheSameTask() throws IOException {
+        Task t1 = new Task("Test1");
+        Task t2 = new Task("Test2");
+
+        list.push(t1);
+        list.push(t2);
+        assertEquals(list.peek(), list.pop());
+    }
+
+    @Test
+    public void popShouldDeleteTask() throws IOException {
+        Task t1 = new Task("Test1");
+        Task t2 = new Task("Test2");
+
+        list.push(t1);
+        list.push(t2);
+        assertEquals(t1, list.pop());
+        assertEquals(t2, list.pop());
+    }
+
+    @Test
+    public void peekShouldLeaveTaskInQueue() throws IOException {
+        Task t1 = new Task("Test1");
+        Task t2 = new Task("Test2");
+
+        list.push(t1);
+        list.push(t2);
+        assertEquals(t1, list.peek());
         assertEquals(t1, list.peek());
     }
 
     @Test
-    public void gettingTaskFromEmptyListShouldThrowException() {
+    public void gettingTaskFromEmptyListShouldThrowException() throws IOException {
         try {
             System.out.println(list.pop());
             list.pop();
